+++ goldilocks/src/poseidon2.rs
@@ -1,6 +1,12 @@
-//! Diffusion matrices for Goldilocks8, Goldilocks12, Goldilocks16, and Goldilocks20.
-//!
-//! Reference: https://github.com/HorizenLabs/poseidon2/blob/main/plain_implementations/src/poseidon2/poseidon2_instance_goldilocks.rs
+//! Implementation of Poseidon2, see: https://eprint.iacr.org/2023/323
+
+//! For now we recreate the implementation given in:
+//! https://github.com/HorizenLabs/poseidon2/blob/main/plain_implementations/src/poseidon2/poseidon2_instance_goldilocks.rs
+//! This uses the constants below along with using the 4x4 matrix:
+//! [[5, 7, 1, 3], [4, 6, 1, 1], [1, 3, 5, 7], [1, 1, 4, 6]]
+//! to build the 4t x 4t matrix used for the external (full) rounds).
+
+//! Long term we will use more optimised internal and external linear layers.
 
 use p3_field::AbstractField;
 use p3_poseidon2::{matmul_internal, DiffusionPermutation};
@@ -120,14 +126,118 @@ impl<AF: AbstractField<F = Goldilocks>> Permutation<[AF; 20]> for DiffusionMatri
 
 impl<AF: AbstractField<F = Goldilocks>> DiffusionPermutation<AF, 20> for DiffusionMatrixGoldilocks {}
 
+pub const HL_GOLDILOCKS_8_EXTERNAL_ROUND_CONSTANTS: [[u64; 8]; 8] = [
+    [
+        0xdd5743e7f2a5a5d9,
+        0xcb3a864e58ada44b,
+        0xffa2449ed32f8cdc,
+        0x42025f65d6bd13ee,
+        0x7889175e25506323,
+        0x34b98bb03d24b737,
+        0xbdcc535ecc4faa2a,
+        0x5b20ad869fc0d033,
+    ],
+    [
+        0xf1dda5b9259dfcb4,
+        0x27515210be112d59,
+        0x4227d1718c766c3f,
+        0x26d333161a5bd794,
+        0x49b938957bf4b026,
+        0x4a56b5938b213669,
+        0x1120426b48c8353d,
+        0x6b323c3f10a56cad,
+    ],
+    [
+        0xce57d6245ddca6b2,
+        0xb1fc8d402bba1eb1,
+        0xb5c5096ca959bd04,
+        0x6db55cd306d31f7f,
+        0xc49d293a81cb9641,
+        0x1ce55a4fe979719f,
+        0xa92e60a9d178a4d1,
+        0x002cc64973bcfd8c,
+    ],
+    [
+        0xcea721cce82fb11b,
+        0xe5b55eb8098ece81,
+        0x4e30525c6f1ddd66,
+        0x43c6702827070987,
+        0xaca68430a7b5762a,
+        0x3674238634df9c93,
+        0x88cee1c825e33433,
+        0xde99ae8d74b57176,
+    ],
+    [
+        0x014ef1197d341346,
+        0x9725e20825d07394,
+        0xfdb25aef2c5bae3b,
+        0xbe5402dc598c971e,
+        0x93a5711f04cdca3d,
+        0xc45a9a5b2f8fb97b,
+        0xfe8946a924933545,
+        0x2af997a27369091c,
+    ],
+    [
+        0xaa62c88e0b294011,
+        0x058eb9d810ce9f74,
+        0xb3cb23eced349ae4,
+        0xa3648177a77b4a84,
+        0x43153d905992d95d,
+        0xf4e2a97cda44aa4b,
+        0x5baa2702b908682f,
+        0x082923bdf4f750d1,
+    ],
+    [
+        0x98ae09a325893803,
+        0xf8a6475077968838,
+        0xceb0735bf00b2c5f,
+        0x0a1a5d953888e072,
+        0x2fcb190489f94475,
+        0xb5be06270dec69fc,
+        0x739cb934b09acf8b,
+        0x537750b75ec7f25b,
+    ],
+    [
+        0xe9dd318bae1f3961,
+        0xf7462137299efe1a,
+        0xb1f6b8eee9adb940,
+        0xbdebcc8a809dfe6b,
+        0x40fc1f791b178113,
+        0x3ac1c3362d014864,
+        0x9a016184bdb8aeba,
+        0x95f2394459fbc25e,
+    ],
+];
+pub const HL_GOLDILOCKS_8_INTERNAL_ROUND_CONSTANTS: [u64; 22] = [
+    0x488897d85ff51f56,
+    0x1140737ccb162218,
+    0xa7eeb9215866ed35,
+    0x9bd2976fee49fcc9,
+    0xc0c8f0de580a3fcc,
+    0x4fb2dae6ee8fc793,
+    0x343a89f35f37395b,
+    0x223b525a77ca72c8,
+    0x56ccb62574aaa918,
+    0xc4d507d8027af9ed,
+    0xa080673cf0b7e95c,
+    0xf0184884eb70dcf8,
+    0x044f10b0cb3d5c69,
+    0xe9e3f7993938f186,
+    0x1b761c80e772f459,
+    0x606cec607a1b5fac,
+    0x14a0c2e1d45f03cd,
+    0x4eace8855398574f,
+    0xf905ca7103eff3e6,
+    0xf8c8f8d20862c059,
+    0xb524fe8bdd678e5a,
+    0xfbb7865901a1ec41,
+];
+
 #[cfg(test)]
 mod tests {
     use core::array;
 
-    use p3_poseidon2::{
-        Poseidon2, HL_GOLDILOCKS_8_EXTERNAL_ROUND_CONSTANTS,
-        HL_GOLDILOCKS_8_INTERNAL_ROUND_CONSTANTS,
-    };
+    use p3_poseidon2::Poseidon2;
 
     use super::*;
