--- poseidon2/src/round_constants.rs
@@ -1,175 +1,0 @@
-// We save the constants used by the Poseidon2 implementation by Horizen Labs:
-// https://github.com/HorizenLabs/poseidon2
-// This lets us test that our implementations matches theirs.
-
-// Note that the Horizen Labs implementation uses the 4x4 matrix:
-// [[5, 7, 1, 3], [4, 6, 1, 1], [1, 3, 5, 7], [1, 1, 4, 6]]
-// to build the 4t x 4t matrix used for the external (full) rounds).
-
-pub const HL_BABYBEAR_16_EXTERNAL_ROUND_CONSTANTS: [[u32; 16]; 8] = [
-    [
-        0x69cbb6af, 0x46ad93f9, 0x60a00f4e, 0x6b1297cd, 0x23189afe, 0x732e7bef, 0x72c246de,
-        0x2c941900, 0x0557eede, 0x1580496f, 0x3a3ea77b, 0x54f3f271, 0x0f49b029, 0x47872fe1,
-        0x221e2e36, 0x1ab7202e,
-    ],
-    [
-        0x487779a6, 0x3851c9d8, 0x38dc17c0, 0x209f8849, 0x268dcee8, 0x350c48da, 0x5b9ad32e,
-        0x0523272b, 0x3f89055b, 0x01e894b2, 0x13ddedde, 0x1b2ef334, 0x7507d8b4, 0x6ceeb94e,
-        0x52eb6ba2, 0x50642905,
-    ],
-    [
-        0x05453f3f, 0x06349efc, 0x6922787c, 0x04bfff9c, 0x768c714a, 0x3e9ff21a, 0x15737c9c,
-        0x2229c807, 0x0d47f88c, 0x097e0ecc, 0x27eadba0, 0x2d7d29e4, 0x3502aaa0, 0x0f475fd7,
-        0x29fbda49, 0x018afffd,
-    ],
-    [
-        0x0315b618, 0x6d4497d1, 0x1b171d9e, 0x52861abd, 0x2e5d0501, 0x3ec8646c, 0x6e5f250a,
-        0x148ae8e6, 0x17f5fa4a, 0x3e66d284, 0x0051aa3b, 0x483f7913, 0x2cfe5f15, 0x023427ca,
-        0x2cc78315, 0x1e36ea47,
-    ],
-    [
-        0x7290a80d, 0x6f7e5329, 0x598ec8a8, 0x76a859a0, 0x6559e868, 0x657b83af, 0x13271d3f,
-        0x1f876063, 0x0aeeae37, 0x706e9ca6, 0x46400cee, 0x72a05c26, 0x2c589c9e, 0x20bd37a7,
-        0x6a2d3d10, 0x20523767,
-    ],
-    [
-        0x5b8fe9c4, 0x2aa501d6, 0x1e01ac3e, 0x1448bc54, 0x5ce5ad1c, 0x4918a14d, 0x2c46a83f,
-        0x4fcf6876, 0x61d8d5c8, 0x6ddf4ff9, 0x11fda4d3, 0x02933a8f, 0x170eaf81, 0x5a9c314f,
-        0x49a12590, 0x35ec52a1,
-    ],
-    [
-        0x58eb1611, 0x5e481e65, 0x367125c9, 0x0eba33ba, 0x1fc28ded, 0x066399ad, 0x0cbec0ea,
-        0x75fd1af0, 0x50f5bf4e, 0x643d5f41, 0x6f4fe718, 0x5b3cbbde, 0x1e3afb3e, 0x296fb027,
-        0x45e1547b, 0x4a8db2ab,
-    ],
-    [
-        0x59986d19, 0x30bcdfa3, 0x1db63932, 0x1d7c2824, 0x53b33681, 0x0673b747, 0x038a98a3,
-        0x2c5bce60, 0x351979cd, 0x5008fb73, 0x547bca78, 0x711af481, 0x3f93bf64, 0x644d987b,
-        0x3c8bcd87, 0x608758b8,
-    ],
-];
-pub const HL_BABYBEAR_16_INTERNAL_ROUND_CONSTANTS: [u32; 13] = [
-    0x5a8053c0, 0x693be639, 0x3858867d, 0x19334f6b, 0x128f0fd8, 0x4e2b1ccb, 0x61210ce0, 0x3c318939,
-    0x0b5b2f22, 0x2edb11d5, 0x213effdf, 0x0cac4606, 0x241af16d,
-];
-pub const HL_BABYBEAR_16_INTERNAL_MAT_DIAG: [u32; 16] = [
-    0x0a632d94, 0x6db657b7, 0x56fbdc9e, 0x052b3d8a, 0x33745201, 0x5c03108c, 0x0beba37b, 0x258c2e8b,
-    0x12029f39, 0x694909ce, 0x6d231724, 0x21c3b222, 0x3c0904a5, 0x01d6acda, 0x27705c83, 0x5231c802,
-];
-
-pub const HL_GOLDILOCKS_8_EXTERNAL_ROUND_CONSTANTS: [[u64; 8]; 8] = [
-    [
-        0xdd5743e7f2a5a5d9,
-        0xcb3a864e58ada44b,
-        0xffa2449ed32f8cdc,
-        0x42025f65d6bd13ee,
-        0x7889175e25506323,
-        0x34b98bb03d24b737,
-        0xbdcc535ecc4faa2a,
-        0x5b20ad869fc0d033,
-    ],
-    [
-        0xf1dda5b9259dfcb4,
-        0x27515210be112d59,
-        0x4227d1718c766c3f,
-        0x26d333161a5bd794,
-        0x49b938957bf4b026,
-        0x4a56b5938b213669,
-        0x1120426b48c8353d,
-        0x6b323c3f10a56cad,
-    ],
-    [
-        0xce57d6245ddca6b2,
-        0xb1fc8d402bba1eb1,
-        0xb5c5096ca959bd04,
-        0x6db55cd306d31f7f,
-        0xc49d293a81cb9641,
-        0x1ce55a4fe979719f,
-        0xa92e60a9d178a4d1,
-        0x002cc64973bcfd8c,
-    ],
-    [
-        0xcea721cce82fb11b,
-        0xe5b55eb8098ece81,
-        0x4e30525c6f1ddd66,
-        0x43c6702827070987,
-        0xaca68430a7b5762a,
-        0x3674238634df9c93,
-        0x88cee1c825e33433,
-        0xde99ae8d74b57176,
-    ],
-    [
-        0x014ef1197d341346,
-        0x9725e20825d07394,
-        0xfdb25aef2c5bae3b,
-        0xbe5402dc598c971e,
-        0x93a5711f04cdca3d,
-        0xc45a9a5b2f8fb97b,
-        0xfe8946a924933545,
-        0x2af997a27369091c,
-    ],
-    [
-        0xaa62c88e0b294011,
-        0x058eb9d810ce9f74,
-        0xb3cb23eced349ae4,
-        0xa3648177a77b4a84,
-        0x43153d905992d95d,
-        0xf4e2a97cda44aa4b,
-        0x5baa2702b908682f,
-        0x082923bdf4f750d1,
-    ],
-    [
-        0x98ae09a325893803,
-        0xf8a6475077968838,
-        0xceb0735bf00b2c5f,
-        0x0a1a5d953888e072,
-        0x2fcb190489f94475,
-        0xb5be06270dec69fc,
-        0x739cb934b09acf8b,
-        0x537750b75ec7f25b,
-    ],
-    [
-        0xe9dd318bae1f3961,
-        0xf7462137299efe1a,
-        0xb1f6b8eee9adb940,
-        0xbdebcc8a809dfe6b,
-        0x40fc1f791b178113,
-        0x3ac1c3362d014864,
-        0x9a016184bdb8aeba,
-        0x95f2394459fbc25e,
-    ],
-];
-pub const HL_GOLDILOCKS_8_INTERNAL_ROUND_CONSTANTS: [u64; 22] = [
-    0x488897d85ff51f56,
-    0x1140737ccb162218,
-    0xa7eeb9215866ed35,
-    0x9bd2976fee49fcc9,
-    0xc0c8f0de580a3fcc,
-    0x4fb2dae6ee8fc793,
-    0x343a89f35f37395b,
-    0x223b525a77ca72c8,
-    0x56ccb62574aaa918,
-    0xc4d507d8027af9ed,
-    0xa080673cf0b7e95c,
-    0xf0184884eb70dcf8,
-    0x044f10b0cb3d5c69,
-    0xe9e3f7993938f186,
-    0x1b761c80e772f459,
-    0x606cec607a1b5fac,
-    0x14a0c2e1d45f03cd,
-    0x4eace8855398574f,
-    0xf905ca7103eff3e6,
-    0xf8c8f8d20862c059,
-    0xb524fe8bdd678e5a,
-    0xfbb7865901a1ec41,
-];
-pub const HL_GOLDILOCKS_8_INTERNAL_MAT_DIAG: [u64; 8] = [
-    0xa98811a1fed4e3a5,
-    0x1cc48b54f377e2a0,
-    0xe40cd4f6c5609a26,
-    0x11de79ebca97a4a3,
-    0x9177c73d8b7e929c,
-    0x2a6fe8085797e791,
-    0x3de6e93329f8d5ad,
-    0x3f7af9125da962fe,
-];
